requirements.txt
# MCP Server
fastmcp>=2.0.0
starlette>=0.27.0
uvicorn>=0.23.0

# Google APIs
google-auth>=2.22.0
google-api-python-client>=2.95.0

# Salesforce
simple-salesforce>=1.12.0

# Environment variables
python-dotenv>=1.0.0

